fix: require cycles
